+++ src/realm_dart.cpp
@@ -96,7 +96,7 @@ RLM_API void realm_dart_invoke_unlock_callback(bool success, void* unlockFunc)
 // Stamped into the library by the build system (see prepare-release.yml)
 // Keep this method as it is written and do not format it. 
 // We have a github workflow that looks for and replaces this string as it is written here.
-RLM_API const char* realm_dart_library_version() { return "0.8.0+rc"; }
+RLM_API const char* realm_dart_library_version() { return "0.9.0+rc"; }
 
 //for debugging only
 // RLM_API void realm_dart_gc() {